Update README.md
The Theme's for catppuccin aur doesn't exist anymore, and I saw you using the kvantum mocha pink theme for catppuccin. So... new package to use for the gtk theme, I think there should be kvantum included in the install packages also, an update to the name of sway-bg to swaybg as thats the new package name / idk if it's a typo on the dev's end.